se cambia nombre en ruta de catalogos-administrativos a configuracion, se valida que no se pueda eliminar empresa y perfil si existen usuarios pertenecientes

+++ src/app/views/catalogos-administrativos/components/empresas/empresas.component.ts
@@ -1,11 +1,12 @@
-import { Component, OnInit,  ViewChild, OnDestroy, ChangeDetectorRef  } from '@angular/core';
+import { Component, OnInit, ViewChild, OnDestroy, ChangeDetectorRef } from '@angular/core';
 import { Empresa } from './../../../../shared/models/empresa';
 import { EmpresasService } from './../../../../shared/services/empresas.service';
-import {MatDialog} from '@angular/material/dialog';
+import { UsuariosService } from './../../../../shared/services/usuarios.service';
+import { MatDialog } from '@angular/material/dialog';
 import { MatTableDataSource, MatPaginator } from '@angular/material';
 import { Observable } from 'rxjs';
 import { Router } from '@angular/router'
-import {MatSnackBar} from '@angular/material/snack-bar';
+import { MatSnackBar } from '@angular/material/snack-bar';
 
 import { ModalEliminarComponent } from '../modal-eliminar/modal-eliminar.component';
 import { environment } from './../../../../../environments/environment';
@@ -21,9 +22,9 @@ export class EmpresasComponent implements OnInit {
   empresasTemp: Empresa[] = [];
   rutaImg: string;
   host: string;
-  
-  
-  @ViewChild(MatPaginator, {static: false}) paginator: MatPaginator;
+
+
+  @ViewChild(MatPaginator, { static: false }) paginator: MatPaginator;
   obs$: Observable<any>;
   dataSource: MatTableDataSource<Empresa> = new MatTableDataSource<Empresa>();
 
@@ -32,7 +33,8 @@ export class EmpresasComponent implements OnInit {
     public dialog: MatDialog,
     private changeDetectorRef: ChangeDetectorRef,
     private router: Router,
-    private snackBar: MatSnackBar
+    private snackBar: MatSnackBar,
+    private usuariosService: UsuariosService
   ) { }
 
   ngOnInit() {
@@ -45,16 +47,16 @@ export class EmpresasComponent implements OnInit {
     this.host = environment.host
   }
 
-  ngOnDestroy(){
-    if (this.dataSource) { 
-      this.dataSource.disconnect(); 
+  ngOnDestroy() {
+    if (this.dataSource) {
+      this.dataSource.disconnect();
     }
   }
 
-  getEmpresas(){
+  getEmpresas() {
     this.empresasService.getAllEmpresas().subscribe(
-      ( (empresas: Empresa[]) => {
-        this.empresas = empresas.filter( empresa => empresa.activo !== 0);
+      ((empresas: Empresa[]) => {
+        this.empresas = empresas.filter(empresa => empresa.activo !== 0);
         console.log(this.empresas);
         this.empresasTemp = this.empresas;
         this.dataSource.data = this.empresas;
@@ -71,7 +73,7 @@ export class EmpresasComponent implements OnInit {
     if (!columns.length)
       return;
 
-    const rows = this.empresasTemp.filter(function(d) {
+    const rows = this.empresasTemp.filter(function (d) {
       for (let i = 0; i <= columns.length; i++) {
         let column = columns[i];
         if (d[column] && d[column].toString().toLowerCase().indexOf(val) > -1) {
@@ -84,39 +86,50 @@ export class EmpresasComponent implements OnInit {
     // console.log(this.dataSource.data);
   }
 
-  openDialoAlertDelete(id) {
+  openDialoAlertDelete(idE) {
     const dialogRef = this.dialog.open(ModalEliminarComponent, {
       width: '300px',
       panelClass: 'custom-dialog-container-delete',
-      data: id
+      data: idE
     });
 
     dialogRef.afterClosed().subscribe(result => {
-      if(result){
+      if (result) {
         const dataEmpresa = {
-          idEmpresa: id,
+          idEmpresa: idE,
           activo: 0
         };
-        console.log(result);
-        console.log(id);
-        this.empresasService.deleteEmpresa(dataEmpresa).subscribe(
-          (response:any) => {
-            if(response.estatus === '05'){
-              this.useAlerts(response.mensaje, ' ', 'success-dialog');
-              this.getEmpresas();
+        this.usuariosService.getUsuarios().subscribe(
+          response => {
+            const usuarios = response.filter(usuario => usuario.perfil.nombre !== 'Bajas' && usuario.idEmpresa === idE);
+            const usuariosPertenecientesEmpresa = usuarios.length;
+            if (usuariosPertenecientesEmpresa === 0) {
+              // ELIMINAR EMPRESA SI NO TIENE USUARIOS
+              this.empresasService.deleteEmpresa(dataEmpresa).subscribe(
+                (response: any) => {
+                  if (response.estatus === '05') {
+                    this.useAlerts(response.mensaje, ' ', 'success-dialog');
+                    this.getEmpresas();
+                  } else {
+                    this.useAlerts(response.mensaje, ' ', 'error-dialog');
+                  }
+                },
+                error => {
+                  console.log(error);
+                }
+              );
             } else {
-              this.useAlerts(response.mensaje, ' ', 'error-dialog');
+              this.useAlerts('No es posible eliminar empresa ya que tiene usuarios activos pertenecientes a la misma', ' ', 'error-dialog');
             }
           },
-          error => {
-            console.log(error);
-          }
+          error => console.log('no se obtuvieron los usuario')
         );
+
       }
     });
   }
 
-  useAlerts(message, action, className){
+  useAlerts(message, action, className) {
     this.snackBar.open(message, action, {
       duration: 4000,
       verticalPosition: 'bottom',
